Created LocalFile model. Updated DropboxFiles scheme. Created DjvuLocalExtension module

+++ lib/pdf_extension.rb
@@ -5,27 +5,31 @@ module PdfExtension
   extend ActiveSupport::Concern
 
   included do
-    before_save :set_book_pages_count, :set_book_cover
+    before_create :set_book_pages_count, :set_book_cover
   end
 
 
   def set_book_pages_count
-    pdf = PDF::Reader.new(pdf_file_io)
-    book.update_column :pages_count, pdf.page_count
+    if pdf_file_io
+      pdf = PDF::Reader.new(pdf_file_io)
+      book.update_column :pages_count, pdf.page_count
+    end
   end
 
 
   #TODO: test deletion after set
   def set_book_cover
-    tmp_bookcover_name = Digest::MD5.hexdigest("#{pdf_file_io.object_id}_cover_#{Time.now.to_i}_#{rand(9999)}")
-    to_delete_filepath = "#{Rails.root}/public/uploads/#{tmp_bookcover_name}.png"
-    im = Magick::Image.read(pdf_file_io.path+"[0]")
-    im[0].write to_delete_filepath
+    if pdf_file_io
+      tmp_bookcover_name = Digest::MD5.hexdigest("#{pdf_file_io.object_id}_cover_#{Time.now.to_i}_#{rand(9999)}")
+      to_delete_filepath = "#{Rails.root}/public/uploads/#{tmp_bookcover_name}.png"
+      im = Magick::Image.read(pdf_file_io.path+"[0]")
+      im[0].write to_delete_filepath
 
-    book.bookcover = File.open(to_delete_filepath)
-    book.save!
+      book.bookcover = File.open(to_delete_filepath)
+      book.save!
 
-    File.delete to_delete_filepath
+      File.delete to_delete_filepath
+    end
   end
